Make Clippy happy again

+++ pyo3-ffi/src/cpython/tupleobject.rs
@@ -24,20 +24,14 @@ pub unsafe fn PyTuple_GET_SIZE(op: *mut PyObject) -> Py_ssize_t {
 #[inline]
 #[cfg(not(PyPy))]
 pub unsafe fn PyTuple_GET_ITEM(op: *mut PyObject, i: Py_ssize_t) -> *mut PyObject {
-    *(*(op as *mut PyTupleObject))
-        .ob_item
-        .as_ptr()
-        .offset(i as isize)
+    *(*(op as *mut PyTupleObject)).ob_item.as_ptr().offset(i)
 }
 
 /// Macro, *only* to be used to fill in brand new tuples
 #[inline]
 #[cfg(not(PyPy))]
 pub unsafe fn PyTuple_SET_ITEM(op: *mut PyObject, i: Py_ssize_t, v: *mut PyObject) {
-    *(*(op as *mut PyTupleObject))
-        .ob_item
-        .as_mut_ptr()
-        .offset(i as isize) = v;
+    *(*(op as *mut PyTupleObject)).ob_item.as_mut_ptr().offset(i) = v;
 }
 
 // skipped _PyTuple_DebugMallocStats
